add registration form page

+++ lib/core/utils/price_formatter.dart
@@ -1,7 +1,7 @@
 import 'package:intl/intl.dart';
 
 class PriceFormatter {
-  static String format(int price) {
+  static String format(double price) {
     final formatter = NumberFormat("#,###", "vi_VN");
     return "${formatter.format(price)}đ";
   }
